requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.1.1
cycler==0.12.1
fonttools==4.54.1
hydra-core==1.3.2
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.41.1
matplotlib==3.7.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
omegaconf==2.3.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tifffile==2023.7.10
zipp==3.20.2
